refactor: extract admin-menu components to sub-components
[#OCD-4307-UI]
