+++ src/app/components/login/admin-menu.jsx
@@ -2,150 +2,21 @@ import React, { useContext, useState } from 'react';
 import {
   Button,
   Box,
-  Collapse,
   List,
-  ListItem,
   makeStyles,
-  Typography,
 } from '@material-ui/core';
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import ExpandLessIcon from '@material-ui/icons/ExpandLess';
 import ExitToAppIcon from '@material-ui/icons/ExitToApp';
 import VpnKeyIcon from '@material-ui/icons/VpnKey';
-import {
-  bool,
-  func,
-  node,
-  shape,
-  string,
-} from 'prop-types';
-import { useCookies } from 'react-cookie';
+import { func } from 'prop-types';
 
-import { usePostLogout } from 'api/auth';
-import { ChplLink } from 'components/util';
+import ChplAdminMenuLinkItem from './navigation/admin-menu-link-item';
+import ChplAdminMenuSection from './navigation/admin-menu-section';
+import sectionConfigs from './navigation/admin-menu-data';
+
 import { getAngularService } from 'services/angular-react-helper';
 import { eventTrack } from 'services/analytics.service';
 import { UserContext, useAnalyticsContext } from 'shared/contexts';
 import { palette } from 'themes';
-
-const sectionConfigs = [{
-  key: 'administration',
-  title: 'Administration',
-  roles: ['chpl-admin', 'chpl-onc', 'chpl-onc-acb', 'chpl-cms-staff'],
-  disablePadding: false,
-  items: [{
-    key: 'upload',
-    roles: ['chpl-admin', 'chpl-onc', 'chpl-onc-acb'],
-    href: '#/administration/upload',
-    text: 'Upload',
-    router: { sref: 'administration.upload' },
-  }, {
-    key: 'confirm-listings',
-    roles: ['chpl-admin', 'chpl-onc-acb'],
-    href: '#/administration/confirm/listings',
-    text: 'Confirm Listings',
-    router: { sref: 'administration.confirm.listings' },
-  }, {
-    key: 'reports',
-    roles: ['chpl-admin', 'chpl-onc', 'chpl-onc-acb'],
-    href: '#/administration/reports',
-    text: 'Reports',
-    router: { sref: 'administration.reports' },
-  }, {
-    key: 'cms',
-    roles: ['chpl-admin', 'chpl-onc', 'chpl-cms-staff'],
-    href: '#/administration/cms',
-    text: 'CMS',
-    router: { sref: 'administration.cms' },
-  }, {
-    key: 'change-requests',
-    roles: ['chpl-admin', 'chpl-onc', 'chpl-onc-acb'],
-    href: '#/administration/change-requests',
-    text: 'Change Requests',
-    router: { sref: 'administration.change-requests' },
-  }, {
-    key: 'system-maintenance',
-    roles: ['chpl-admin', 'chpl-onc', 'chpl-onc-acb'],
-    href: '#/administration/system-maintenance',
-    text: 'System Maintenance',
-    router: { sref: 'administration.system-maintenance' },
-  }, {
-    key: 'url-checker',
-    roles: ['chpl-admin', 'chpl-onc'],
-    href: '#/administration/url-checker',
-    text: 'URL Checker',
-    router: { sref: 'administration.url-checker' },
-  }, {
-    key: 'user-management',
-    roles: ['chpl-admin', 'chpl-onc'],
-    href: '#/users',
-    text: 'User Management',
-    router: { sref: 'users' },
-  }, {
-    key: 'ff4j',
-    roles: ['chpl-admin'],
-    href: 'rest/ff4j-console/home',
-    text: 'FF4j',
-    external: true,
-  }],
-}, {
-  key: 'organizations',
-  title: 'Organizations',
-  roles: ['chpl-admin', 'chpl-onc', 'chpl-onc-acb'],
-  items: [{
-    key: 'developers',
-    href: '#/organizations/developers',
-    text: 'Developers',
-    router: { sref: 'organizations.developers' },
-  }, {
-    key: 'onc-acbs',
-    href: '#/organizations/onc-acbs',
-    text: 'ONC-ACBs',
-    router: { sref: 'organizations.onc-acbs' },
-  }, {
-    key: 'onc-atls',
-    roles: ['chpl-admin', 'chpl-onc'],
-    href: '#/organizations/onc-atls',
-    text: 'ONC-ATLs',
-    router: { sref: 'organizations.onc-atls' },
-  }],
-}, {
-  key: 'activity',
-  title: 'Activity',
-  roles: ['chpl-admin', 'chpl-onc'],
-  items: [{
-    key: 'activity-search',
-    href: '#/reports/activity',
-    text: 'Search',
-    router: { sref: 'reports.activity' },
-  }, {
-    key: 'questionable-activity',
-    href: '#/reports/questionable-activity',
-    text: 'Questionable Activity',
-    router: { sref: 'reports.questionable-activity' },
-  }],
-}, {
-  key: 'surveillance',
-  title: 'Surveillance',
-  roles: ['chpl-admin', 'chpl-onc', 'chpl-onc-acb'],
-  items: [{
-    key: 'activity-reporting',
-    roles: ['chpl-admin', 'chpl-onc'],
-    href: '#/surveillance/activity-reporting',
-    text: 'Activity Reporting',
-    router: { sref: 'surveillance.activity-reporting' },
-  }, {
-    key: 'complaints-reporting',
-    href: '#/surveillance/complaints',
-    text: 'Complaints Reporting',
-    router: { sref: 'surveillance.complaints' },
-  }, {
-    key: 'reporting',
-    href: '#/surveillance/reporting',
-    text: 'Reporting',
-    router: { sref: 'surveillance.reporting' },
-  }],
-}];
 
 const useStyles = makeStyles({
   menuContainer: {
@@ -154,43 +25,6 @@ const useStyles = makeStyles({
     width: '280px',
     color: palette.greyDark,
     backgroundColor: palette.white,
-  },
-  sectionHeader: {
-    padding: '12px 16px',
-    cursor: 'pointer',
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    color: palette.greyDark,
-    '&:hover': {
-      backgroundColor: palette.secondary,
-    },
-  },
-  sectionHeaderText: {
-    fontWeight: 500,
-    fontSize: '16px',
-    display: 'flex',
-    flexGrow: 1,
-  },
-  sectionChevron: {
-    marginLeft: 'auto',
-  },
-  menuItem: {
-    padding: '8px 16px 8px 32px',
-    cursor: 'pointer',
-    color: palette.primary,
-    fontSize: '14px',
-    textDecoration: 'underline',
-    '& a': {
-      color: `${palette.primary}!important`,
-      textDecoration: 'underline!important',
-    },
-    '& .MuiSvgIcon-root': {
-      color: palette.primary,
-    },
-    '&:hover': {
-      backgroundColor: palette.secondary,
-    },
   },
   footer: {
     display: 'flex',
@@ -201,91 +35,13 @@ const useStyles = makeStyles({
   },
 });
 
-function AdminMenuSection({
-  children,
-  classes,
-  disablePadding = true,
-  isOpen,
-  onToggle,
-  section,
-  title,
-}) {
-  return (
-    <>
-      <ListItem
-        className={classes.sectionHeader}
-        onClick={() => onToggle(section)}
-        divider
-      >
-        <Typography className={classes.sectionHeaderText}>{title}</Typography>
-        {isOpen
-          ? <ExpandLessIcon className={classes.sectionChevron} />
-          : <ExpandMoreIcon className={classes.sectionChevron} />}
-      </ListItem>
-      <Collapse in={isOpen} timeout="auto" unmountOnExit>
-        <List component="div" disablePadding={disablePadding}>
-          {children}
-        </List>
-      </Collapse>
-    </>
-  );
-}
-
-AdminMenuSection.propTypes = {
-  children: node.isRequired,
-  classes: shape({}).isRequired,
-  disablePadding: bool,
-  isOpen: bool.isRequired,
-  onToggle: func.isRequired,
-  section: string.isRequired,
-  title: string.isRequired,
-};
-
-function AdminMenuLinkItem({
-  classes,
-  external = false,
-  href,
-  onClose,
-  router = undefined,
-  text,
-}) {
-  return (
-    <ListItem className={classes.menuItem} onClick={onClose}>
-      <ChplLink
-        href={href}
-        text={text}
-        external={external}
-        router={router}
-      />
-    </ListItem>
-  );
-}
-
-AdminMenuLinkItem.propTypes = {
-  classes: shape({
-    menuItem: string,
-  }).isRequired,
-  external: bool,
-  href: string.isRequired,
-  onClose: func.isRequired,
-  router: shape({
-    sref: string,
-  }),
-  text: string.isRequired,
-};
-
 function ChplAdminMenu({ onClose = () => {} }) {
-  const $rootScope = getAngularService('$rootScope');
   const authService = getAngularService('authService');
-  const [, , removeCookie] = useCookies(['cognito_id', 'refresh_token']);
   const { analytics } = useAnalyticsContext();
   const {
     hasAnyRole,
     setLoginWidgetState,
-    user,
-    setUser,
   } = useContext(UserContext);
-  const postLogout = usePostLogout();
   const [openSection, setOpenSection] = useState(null);
   const classes = useStyles();
 
@@ -322,9 +78,8 @@ function ChplAdminMenu({ onClose = () => {} }) {
         { sectionConfigs
           .filter((sectionConfig) => canAccess(sectionConfig.roles))
           .map((sectionConfig) => (
-            <AdminMenuSection
+            <ChplAdminMenuSection
               key={sectionConfig.key}
-              classes={classes}
               section={sectionConfig.key}
               title={sectionConfig.title}
               isOpen={openSection === sectionConfig.key}
@@ -334,9 +89,8 @@ function ChplAdminMenu({ onClose = () => {} }) {
               { sectionConfig.items
                 .filter((item) => canAccess(item.roles))
                 .map((item) => (
-                  <AdminMenuLinkItem
+                  <ChplAdminMenuLinkItem
                     key={item.key}
-                    classes={classes}
                     href={item.href}
                     text={item.text}
                     external={item.external}
@@ -344,7 +98,7 @@ function ChplAdminMenu({ onClose = () => {} }) {
                     onClose={onClose}
                   />
                 ))}
-            </AdminMenuSection>
+            </ChplAdminMenuSection>
           ))}
       </List>
       <Box className={classes.footer}>
